+++ lib/main_db/core/triggers/documents_triggers.dart
@@ -73,8 +73,7 @@ OLD.icon_value,
         page_count
       )
       SELECT
-        (SELECT id FROM vault_item_history
-         WHERE item_id = OLD.id ORDER BY action_at DESC LIMIT 1),
+        (SELECT id FROM vault_item_history WHERE rowid = last_insert_rowid()),
         di.document_type,
         di.aggregated_text,
         di.aggregate_hash,
@@ -150,8 +149,7 @@ v.icon_value,
         aggregate_hash,
         page_count
       ) VALUES (
-        (SELECT id FROM vault_item_history
-         WHERE item_id = OLD.item_id ORDER BY action_at DESC LIMIT 1),
+        (SELECT id FROM vault_item_history WHERE rowid = last_insert_rowid()),
         OLD.document_type,
         OLD.aggregated_text,
         OLD.aggregate_hash,
@@ -218,8 +216,7 @@ OLD.icon_value,
         page_count
       )
       SELECT
-        (SELECT id FROM vault_item_history
-         WHERE item_id = OLD.id ORDER BY action_at DESC LIMIT 1),
+        (SELECT id FROM vault_item_history WHERE rowid = last_insert_rowid()),
         di.document_type,
         di.aggregated_text,
         di.aggregate_hash,